Added additional information for libraries
When testing a library, it seem to be necessary to add the `--lib` flag to cargo for complication to even succeed.
When using the example provided in the README.md file, in `src/main.rs`, no problems arise, but renaming the same file to `src/lib.rs` render the rustc unable to compile the library for testing, both versions using `cargo test`. Adding the `--lib` flag fixes this problem for libraries.
I assume this is not necessarily your fault, but it is worth providing users with the information.